FISH-868 JWT public key is cached for a configurable amount of time.
